+++ src/server/scripts/BrawlersGuild/brawlers_guild_bosses_rank_four.cpp
@@ -1,3 +1,20 @@
+/*
+ * This file is part of the DestinyCore Project. See AUTHORS file for Copyright information
+ *
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the
+ * Free Software Foundation; either version 2 of the License, or (at your
+ * option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
 #include "AchievementMgr.h"
 #include "BrawlersGuild.h"
 #include "SpellScript.h"
@@ -157,7 +174,7 @@ public:
                 {
                     case EVENT_1:
                     Position pos;
-                    me->GetNearPoint2D(pos, 15.0f, frand(-M_PI/3, M_PI/3));
+                    me->GetNearPoint2D(pos, 15.0f, frand(float(- M_PI / 3), float(M_PI / 3)));
                     me->CastSpell(pos.GetPositionX(), pos.GetPositionY(), pos.GetPositionZ(), SPELL_WATER_ORB);
                     events.RescheduleEvent(EVENT_1, urand (17000, 19000));
                     break;
