+++ src/server/scripts/Draenor/Highmaul/boss_the_butcher.cpp
@@ -1,10 +1,19 @@
-////////////////////////////////////////////////////////////////////////////////
-//
-//  MILLENIUM-STUDIO
-//  Copyright 2016 Millenium-studio SARL
-//  All Rights Reserved.
-//
-////////////////////////////////////////////////////////////////////////////////
+/*
+ * This file is part of the DestinyCore Project. See AUTHORS file for Copyright information
+ *
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the
+ * Free Software Foundation; either version 2 of the License, or (at your
+ * option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
 
 # include "highmaul.hpp"
 #include "SpellAuraEffects.h"
@@ -494,7 +503,7 @@ struct boss_the_butcher : public BossAI
 
                 for (uint8 i = 0; i < count; ++i)
                 {
-                    float o = frand(0, 2 * M_PI);
+                    float o = frand(0, float(2 * M_PI));
                     float x = posX + (radius * cos(o));
                     float y = posY + (radius * sin(o));
