+++ src/server/scripts/Draenor/Highmaul/boss_twin_ogron.cpp
@@ -1,10 +1,19 @@
-////////////////////////////////////////////////////////////////////////////////
-//
-//  MILLENIUM-STUDIO
-//  Copyright 2016 Millenium-studio SARL
-//  All Rights Reserved.
-//
-////////////////////////////////////////////////////////////////////////////////
+/*
+ * This file is part of the DestinyCore Project. See AUTHORS file for Copyright information
+ *
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the
+ * Free Software Foundation; either version 2 of the License, or (at your
+ * option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
 
 # include "highmaul.hpp"
 #include "SpellAuraEffects.h"
@@ -588,7 +597,7 @@ struct boss_twin_ogron_phemos : public BossAI
                 m_EnfeeblingRoarScheduled = false;
                 m_WhirlwindScheduled = false;
 
-                float o = frand(0, 2 * M_PI);
+                float o = frand(0, float(2 * M_PI));
                 Position pos = {(centerPos.m_positionX + (circleToCenterDist * cos(o))), (centerPos.m_positionY + (circleToCenterDist * sin(o))), centerPos.m_positionZ, 0.0f};
 
                 pos.m_orientation = pos.GetAngle(&centerPos);
@@ -604,7 +613,7 @@ struct boss_twin_ogron_phemos : public BossAI
                 me->CastSpell(pos, BlazeWeaponVisual, true);
                 me->SendPlaySpellVisualKit(QuakeVisualID, 0, 0);
 
-                o += M_PI;
+                o += float(M_PI);
                 pos = {(centerPos.m_positionX + (circleToCenterDist * cos(o))), (centerPos.m_positionY + (circleToCenterDist * sin(o))), centerPos.m_positionZ, 0.0f};
 
                 pos.m_orientation = pos.GetAngle(&centerPos);
@@ -915,7 +924,7 @@ struct boss_twin_ogron_phemos : public BossAI
     {
         if (SpellInfo const* spellInfo = sSpellMgr->GetSpellInfo(BlazeFirstSpawn))
         {
-            float oStep = (pos.m_orientation - (M_PI / 2.0f)) + frand(0.0f, (M_PI / 3.0f));
+            float oStep = (pos.m_orientation - (M_PI / 2.0f)) + frand(0.0f, float(M_PI / 3.0f));
 
             for (uint8 i = 0; i < BlazeFirstSpawnCounter; ++i)
             {
@@ -953,7 +962,7 @@ struct boss_twin_ogron_phemos : public BossAI
                     timer += 500;
                 }
 
-                oStep += M_PI / 3.0f;
+                oStep += float(M_PI / 3.0f);
             }
         }
     }
