+++ src/server/scripts/EasternKingdoms/ShadowfangKeep/boss_lord_godfrey.cpp
@@ -1,3 +1,20 @@
+/*
+ * This file is part of the DestinyCore Project. See AUTHORS file for Copyright information
+ *
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the
+ * Free Software Foundation; either version 2 of the License, or (at your
+ * option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
 #include"Spell.h"
 #include"shadowfang_keep.h"
 
@@ -230,7 +247,7 @@ class spell_pistol_barrage : public SpellScriptLoader
 
             void OnPeriodic(AuraEffect const* aurEff)
             {
-                float angle = frand(0, M_PI / 6);
+                float angle = frand(0, float(M_PI / 6));
                 Position pos1;
                 Position pos2;
                 GetCaster()->GetNearPosition(pos1, 45.0f, angle);
