+++ src/server/scripts/Kalimdor/CavernsOfTime/DragonSoul/boss_morchok.cpp
@@ -1,3 +1,20 @@
+/*
+ * This file is part of the DestinyCore Project. See AUTHORS file for Copyright information
+ *
+ * This program is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License as published by the
+ * Free Software Foundation; either version 2 of the License, or (at your
+ * option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
+ * more details.
+ *
+ * You should have received a copy of the GNU General Public License along
+ * with this program. If not, see <http://www.gnu.org/licenses/>.
+ */
+
 #include "dragon_soul.h"
 
 enum ScriptedTexts
@@ -358,7 +375,7 @@ class boss_morchok: public CreatureScript
                             Talk(ANN_CRYSTAL);
                             Talk(SAY_CRYSTAL);
                             Position pos;
-                            me->GetNearPosition(pos, frand(25.0f, 45.0f), frand(0, 2 * M_PI));
+                            me->GetNearPosition(pos, frand(25.0f, 45.0f), frand(0, float(2 * M_PI)));
                             me->CastSpell(pos.GetPositionX(), pos.GetPositionY(), pos.GetPositionZ(), SPELL_RESONATING_CRYSTAL, true);
                             events.ScheduleEvent(EVENT_RESONATING_CRYSTAL, urand(12000, 14000));
                             if (bKohcrom && !bFirstCrystal && pKohcrom)
@@ -543,7 +560,7 @@ class npc_morchok_kohcrom: public CreatureScript
                         case EVENT_RESONATING_CRYSTAL:
                         {
                             Position pos;
-                            me->GetNearPosition(pos, frand(25.0f, 45.0f), frand(0, 2 * M_PI));
+                            me->GetNearPosition(pos, frand(25.0f, 45.0f), frand(0, float(2 * M_PI)));
                             me->CastSpell(pos.GetPositionX(), pos.GetPositionY(), pos.GetPositionZ(), SPELL_RESONATING_CRYSTAL, true);
                             break;
                         }
